feat: prettify help screen
Also make it so that scrolling for the diagnostics view is paused while
the help screen is visible.
