+++ tricorder.cabal
@@ -41,6 +41,7 @@ library
       Tricorder.SourceLookup
       Tricorder.UI
       Tricorder.UI.Event
+      Tricorder.UI.Misc
       Tricorder.UI.State
       Tricorder.UI.View
       Tricorder.Watcher
